storage_service: Remove _is_bootstrap_mode
This "state" is the sub-state of the STARTING mode that's activated
when the storage_service::bootstrap() is called. Instead of the
separate boolean the new mode can be used. To stop it from reverting
the BOOTSTRAP mode back to JOINING some calls to set_mode() should
be converted into regular logging.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>