compaction_manager: let caller serialize maintenance operations per keyspace
Maintenance operations need not be serialized globally.
There is no reason not to run e.g. cleanup or reshape on
multiple keyspaces and even multiple tables within
each keyspace in parallel, so let the api layer
serialize maintenance operations per keyspace
and schedule the compaction_manager tasks in the keyspace.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>