treewide: fix compilation issues with fmtlib 8.1.0+
Due to https://github.com/fmtlib/fmt/commit/fd62fba9855c9cf8fe17eb9363e03f93563d9dc3
scoped enums are not automatically converted to integers anymore,
this is the intended behavior, according to the fmtlib devs.

A bit nicer solution would be to use `std::to_underlying`
instead of a direct `static_cast`, but it's not available until
C++23 and some compilers are still missing the support for it.

Tests: unit(dev)

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>