db,docs: set alternator default write policy
This commit sets the default isolation policy for alternator.
Previously, users were required to explicitly choose their isolation
policy, but that adds another step to the quickstart process
and can slow the actual adoption, as users might find the choice
to be too complicated to bother.
The chosen policy is always_use_lwt, which induces a performance
penalty, but is also safest out of all isolation policies, allowing
the users to mix conditional and unconditional updates.