property-file snitch: Use _my_distributed to re-shard
The driver in question wants to execute some of its actions on shard 0
and it calls smp::invoke(0, ...) for this. The invoked lambda thus needs
to refer to global snitch instance.

There's nicer and shorter way of re-sharding for snith drivers -- the
sharded<snith_ptr>* _my_distributed field on the base class.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>