property-file snitch: Re-gossip states with the help of .get_app_states()
This is the last place that still uses gossip_snitch_info(). It
can be reworked to use the get_app_states(), then the former
helper can be removed.

Another motivation for this is to stop using the _gossiper_started
boolean from the base class. This, in turn, will allow to remove
the whole gossiper_starting() notification altogether.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>